Renaming model to be just UpgradeResult. Nesting the enum in the same class. Introducing a new Controller for the  Install endpoint
