+++ OurUmbraco/Our/Api/UpgradeCheckController.cs
@@ -1,6 +1,4 @@
 ﻿using System.Linq;
-using System.Net;
-using System.Net.Http;
 using System.Web;
 using System.Web.Http;
 using System.Web.Security;
@@ -9,108 +7,12 @@ using OurUmbraco.Our.Models;
 using OurUmbraco.Our.Services;
 using Umbraco.Core.Persistence;
 using Umbraco.Web.WebApi;
+using static OurUmbraco.Our.Models.UpgradeResult;
 
 namespace OurUmbraco.Our.Api
 {
     public class UpgradeCheckController : UmbracoApiController
     {
-        [HttpPost]
-        public IHttpActionResult Install(InstallationModel model)
-        {
-            if (!ModelState.IsValid)
-                throw new HttpResponseException(Request.CreateValidationErrorResponse(ModelState));
-
-            var hasError = !string.IsNullOrEmpty(model.Error);
-            var ipAddress = HttpContext.Current.Request.UserHostAddress;
-            var domainName = HttpContext.Current.Request.UserHostName;
-
-            if (string.IsNullOrWhiteSpace(model.VersionComment) == false)
-            {
-                // V8 nightlies are reporting versions like alpha.58.1927 - which is too long,
-                // the column in the database is only 10 chars. Abbreviating the version comment here will result in fewer 
-                // errors logged in people's installs and we can see how many people actually 
-                // try the nightlies
-                if (model.VersionComment.Length > 10 && model.VersionComment.Contains("alpha"))
-                {
-                    // V8 nightlies all start with "alpha"
-                    model.VersionComment = model.VersionComment.Replace("alpha", "a");
-                }
-                else
-                {
-                    if (model.VersionComment.Length > 0 && model.VersionComment.Length > 10)
-                        // truncate to 10 chars, the max size
-                        model.VersionComment = model.VersionComment.Substring(0, 10);
-                }
-            }
-
-            if (!model.InstallCompleted)
-            {
-                using (var umbracoUpdateDb = new Database("umbracoUpdate"))
-                {
-                    var insertCmd = @"INSERT INTO [installs]
-           ([completed]
-           ,[isUpgrade]
-           ,[errorLogged]
-           ,[installId]
-           ,[installStart]
-           ,[versionMajor]
-           ,[versionMinor]
-           ,[versionPatch]
-           ,[versionComment]
-           ,[ip]
-           ,[domain]
-           ,[userAgent])
-     VALUES
-           (@completed
-           ,@isUpgrade
-           ,@errorLogged
-           ,@installId
-           ,@installStart
-           ,@versionMajor
-           ,@versionMinor
-           ,@versionPatch
-           ,@versionComment
-           ,@ip
-           ,@domain
-           ,@userAgent)";
-
-                    umbracoUpdateDb.Execute(insertCmd, new
-                    {
-                        completed = model.InstallCompleted,
-                        isUpgrade = model.IsUpgrade,
-                        errorLogged = hasError,
-                        installId = model.InstallId,
-                        installStart = model.Timestamp,
-                        versionMajor = model.VersionMajor,
-                        versionMinor = model.VersionMinor,
-                        versionPatch = model.VersionPatch,
-                        versionComment = model.VersionComment,
-                        ip = ipAddress,
-                        domain = domainName,
-                        userAgent = model.UserAgent
-                    });
-                }
-            }
-            else
-            {
-                using (var umbracoUpdateDb = new Database("umbracoUpdate"))
-                {
-                    var updateCmd = @"UPDATE [installs]
-           SET [completed] = 1, [installEnd] = @installEnd, [dbProvider] = @dbProvider 
-            WHERE installId = @installId";
-
-                    umbracoUpdateDb.Execute(updateCmd, new
-                    {
-                        installId = model.InstallId,
-                        installEnd = model.Timestamp,
-                        dbProvider = model.DbProvider
-                    });
-                }
-            }
-
-            return Ok();
-        }
-
         [HttpPost]
         public IHttpActionResult CheckUpgrade(UpgradeModel model)
         {
@@ -123,7 +25,7 @@ namespace OurUmbraco.Our.Api
 
             var latestVersion = GetLatestVersionFromMajor(model.VersionMajor);
             if (latestVersion == null)
-                return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.None, "", "")));
+                return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.None, "", "")));
 
             // Persist the update check for our statistics, don't remove!
             var caller = HttpContext.Current.Request.UserHostName + "_" +
@@ -135,40 +37,40 @@ namespace OurUmbraco.Our.Api
 
             // Special case for 4.0.4.2, apperently we never recommended them to upgrade
             if (version == new System.Version(4, 0, 4, 2) || version == latestVersion)
-                return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.None, "", "")));
+                return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.None, "", "")));
 
             if (version.Major == 4)
             {
                 // We had some special cases in the old updatechecker so I left them here
                 if (version == new System.Version(4, 0, 4, 1))
-                    return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.Minor, "4.0.4.2 is out with fixes for load-balanced sites.", "http://our.umbraco.org/download")));
+                    return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.Minor, "4.0.4.2 is out with fixes for load-balanced sites.", "http://our.umbraco.org/download")));
 
                 if (version == new System.Version(4, 0, 4, 0))
-                    return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.Critical, "4.0.4.1 is out fixing a serious macro bug. Please upgrade!", "http://our.umbraco.org/download")));
+                    return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.Critical, "4.0.4.1 is out fixing a serious macro bug. Please upgrade!", "http://our.umbraco.org/download")));
 
 
                 if (version == new System.Version(4, 7, 0, 0))
-                    return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.Critical, "This Umbraco installation needs to be upgraded. It may contain a potential security issue!", "http://our.umbraco.org/download")));
+                    return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.Critical, "This Umbraco installation needs to be upgraded. It may contain a potential security issue!", "http://our.umbraco.org/download")));
 
 
                 if (version >= new System.Version(4, 10, 0, 0) && version < latestVersion)
-                    return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.Major, $"{latestVersion} is released. Upgrade today - it's free!", "http://our.umbraco.org/download")));
+                    return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.Major, $"{latestVersion} is released. Upgrade today - it's free!", "http://our.umbraco.org/download")));
             }
 
             if (version.Major == 6)
             {
                 if ((version < latestVersion))
-                    return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.Major, $"{latestVersion} is released. Upgrade today - it's free!", "http://our.umbraco.org/download")));
+                    return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.Major, $"{latestVersion} is released. Upgrade today - it's free!", "http://our.umbraco.org/download")));
             }
 
             if (version.Major == 7 || version.Major == 8 || version.Major > 8)
             {
                 if (version < latestVersion)
-                    return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.Minor, $"{latestVersion} is released. Upgrade today - it's free!", $"http://our.umbraco.org/contribute/releases/{latestVersion.Major}{latestVersion.Minor}{latestVersion.Build}")));
+                    return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.Minor, $"{latestVersion} is released. Upgrade today - it's free!", $"http://our.umbraco.org/contribute/releases/{latestVersion.Major}{latestVersion.Minor}{latestVersion.Build}")));
             }
 
             // If nothing matches then it's probably a nightly or a very old version, no need to send upgrade message
-            return Ok(JsonConvert.SerializeObject(new UpgradeResultModel(UpgradeType.None, "", "")));
+            return Ok(JsonConvert.SerializeObject(new UpgradeResult(UpgradeType.None, "", "")));
         }
 
         private void PersistUpdateCheck(string server, int majorVersion, int minorVersion, int patchVersion, string commentVersion)
