+++ OurUmbraco/Our/Models/UpgradeResult.cs
@@ -3,30 +3,30 @@ using Newtonsoft.Json.Converters;
 
 namespace OurUmbraco.Our.Models
 {
-    public enum UpgradeType
+    public class UpgradeResult
     {
-        None,
-        Patch,
-        Minor,
-        Major,
-        Critical,
-        Error,
-        OutOfSync
-    }
+        public enum UpgradeType
+        {
+            None,
+            Patch,
+            Minor,
+            Major,
+            Critical,
+            Error,
+            OutOfSync
+        }
 
-    public class UpgradeResultModel
-    {
         [JsonProperty(PropertyName = "upgradeType")]
         [JsonConverter(typeof(StringEnumConverter))]
-        public UpgradeType UpgradeType { get; set; }
+        public UpgradeType CurrentUpgradeType { get; set; }
         [JsonProperty(PropertyName = "comment")]
         public string Comment { get; set; }
         [JsonProperty(PropertyName = "upgradeUrl")]
         public string UpgradeUrl { get; set; }
 
-        public UpgradeResultModel(UpgradeType upgradeType, string comment, string upgradeUrl)
+        public UpgradeResult(UpgradeType upgradeType, string comment, string upgradeUrl)
         {
-            UpgradeType = upgradeType;
+            CurrentUpgradeType = upgradeType;
             Comment = comment;
             UpgradeUrl = upgradeUrl;
         }
