+++ OurUmbraco/OurUmbraco.csproj
@@ -488,6 +488,7 @@
     <Compile Include="Location\LocationService.cs" />
     <Compile Include="Location\Models\Location.cs" />
     <Compile Include="Our\Api\DashboardStatisticsController.cs" />
+    <Compile Include="Our\Api\InstallationController.cs" />
     <Compile Include="Our\Api\MembersController.cs" />
     <Compile Include="Our\Api\OurMemberStatisticsController.cs" />
     <Compile Include="Our\Api\IssuesStatisticsController.cs" />
@@ -565,7 +566,7 @@
     <Compile Include="Our\Models\Releases.cs" />
     <Compile Include="Community\Models\TeamUmbraco.cs" />
     <Compile Include="Our\Models\UpgradeModel.cs" />
-    <Compile Include="Our\Models\UpgradeResultModel.cs" />
+    <Compile Include="Our\Models\UpgradeResult.cs" />
     <Compile Include="Our\Services\ContributorBadgeService.cs" />
     <Compile Include="Our\Services\EmailService.cs" />
     <Compile Include="Our\Models\ActivationEmailModel.cs" />
